+++ packages/frontend/src/features/semanticViewer/store/cartesianChartBaseSlice.ts
@@ -1,6 +1,7 @@
 import {
     DEFAULT_AGGREGATION,
     IndexType,
+    isFormat,
     type AggregationOptions,
     type BarChartSqlConfig,
     type CartesianChartDisplay,
@@ -17,9 +18,9 @@ type InitialState = {
     defaultLayout: SqlCartesianChartLayout | undefined;
     config: BarChartSqlConfig | LineChartSqlConfig | undefined;
     options: {
-        xLayoutOptions: IndexLayoutOptions[];
-        yLayoutOptions: ValuesLayoutOptions[];
-        groupByOptions: PivotLayoutOptions[];
+        indexLayoutOptions: IndexLayoutOptions[];
+        valuesLayoutOptions: ValuesLayoutOptions[];
+        pivotLayoutOptions: PivotLayoutOptions[];
     };
 };
 
@@ -27,9 +28,9 @@ const initialState: InitialState = {
     defaultLayout: undefined,
     config: undefined,
     options: {
-        xLayoutOptions: [],
-        yLayoutOptions: [],
-        groupByOptions: [],
+        indexLayoutOptions: [],
+        valuesLayoutOptions: [],
+        pivotLayoutOptions: [],
     },
 };
 
@@ -44,7 +45,7 @@ export const cartesianChartConfigSlice = createSlice({
             if (state.config?.fieldConfig?.x) {
                 state.config.fieldConfig.x.reference = action.payload;
                 state.config.fieldConfig.x.type =
-                    state.options.xLayoutOptions.find(
+                    state.options.indexLayoutOptions.find(
                         (x) => x.reference === action.payload,
                     )?.type ?? IndexType.CATEGORY;
             }
@@ -78,7 +79,7 @@ export const cartesianChartConfigSlice = createSlice({
                 if (yAxis) {
                     yAxis.reference = action.payload.reference;
                     yAxis.aggregation =
-                        state.options.yLayoutOptions.find(
+                        state.options.valuesLayoutOptions.find(
                             (option) =>
                                 option.reference === action.payload.reference,
                         )?.aggregationOptions[0] ?? DEFAULT_AGGREGATION;
@@ -170,12 +171,13 @@ export const cartesianChartConfigSlice = createSlice({
             if (!state.config) return;
             if (!state.config.fieldConfig) return;
 
-            const yAxisFieldsAvailable = state.options.yLayoutOptions.filter(
-                (option) =>
-                    !state.config?.fieldConfig?.y
-                        .map((y) => y.reference)
-                        .includes(option.reference),
-            );
+            const yAxisFieldsAvailable =
+                state.options.valuesLayoutOptions.filter(
+                    (option) =>
+                        !state.config?.fieldConfig?.y
+                            .map((y) => y.reference)
+                            .includes(option.reference),
+                );
             const yAxisFields = state.config.fieldConfig.y;
 
             let defaultYAxisField: string | undefined;
@@ -194,10 +196,104 @@ export const cartesianChartConfigSlice = createSlice({
             }
         },
         removeYAxisField: (state, action: PayloadAction<number>) => {
-            if (!state.config) return;
-            if (!state.config.fieldConfig) return;
+            if (!state.config?.fieldConfig?.y) return;
 
-            state.config.fieldConfig.y.splice(action.payload, 1);
+            const index = action.payload;
+            state.config.fieldConfig.y.splice(index, 1);
+
+            if (state.config.display) {
+                if (state.config.display.yAxis) {
+                    state.config.display.yAxis.splice(index, 1);
+                }
+
+                if (state.config.display.series) {
+                    Object.entries(state.config.display.series).forEach(
+                        ([key, series]) => {
+                            if (series.yAxisIndex === index) {
+                                // NOTE: Delete series from display object when it's removed from yAxis
+                                delete state.config?.display?.series?.[key];
+                            } else if (
+                                // NOTE: Decrease yAxisIndex for series that are after the removed yAxis
+                                series.yAxisIndex &&
+                                series.yAxisIndex > index
+                            ) {
+                                series.yAxisIndex--;
+                            }
+                        },
+                    );
+                }
+            }
+        },
+        setStacked: ({ config }, action: PayloadAction<boolean>) => {
+            if (!config) return;
+
+            config.display = config.display || {};
+
+            config.display.stack = action.payload;
+        },
+
+        setYAxisFormat: (
+            { config },
+            action: PayloadAction<{ format: string }>,
+        ) => {
+            if (!config) return;
+            config.display = config.display || {};
+
+            const validFormat = isFormat(action.payload.format)
+                ? action.payload.format
+                : undefined;
+
+            config.display.yAxis = config.display.yAxis || [];
+
+            if (config.display.yAxis.length === 0) {
+                config.display.yAxis.push({ format: validFormat });
+            } else {
+                config.display.yAxis[0].format = validFormat;
+            }
+
+            // Update the format for series with yAxisIndex 0
+            if (config.display.series) {
+                Object.values(config.display.series).forEach((series) => {
+                    if (series.yAxisIndex === 0) {
+                        series.format = validFormat;
+                    }
+                });
+            } else if (config.fieldConfig?.y[0].reference) {
+                config.display.series = {
+                    [config.fieldConfig?.y[0].reference]: {
+                        format: validFormat,
+                        yAxisIndex: 0,
+                    },
+                };
+            }
+        },
+        setSeriesFormat: (
+            { config },
+            action: PayloadAction<{
+                index: number;
+                format: string;
+                reference: string;
+            }>,
+        ) => {
+            if (!config?.display) return;
+
+            const { index, format, reference } = action.payload;
+            const validFormat = isFormat(format) ? format : undefined;
+
+            config.display.series = config.display.series || {};
+            config.display.series[reference] = {
+                ...config.display.series[reference],
+                format: validFormat,
+                yAxisIndex: index,
+            };
+
+            if (index === 0) {
+                config.display.yAxis = config.display.yAxis || [];
+                config.display.yAxis[0] = {
+                    ...config.display.yAxis[0],
+                    format: validFormat,
+                };
+            }
         },
     },
 });