+++ packages/frontend/src/features/sqlRunner/components/Sidebar.tsx
@@ -2,15 +2,17 @@ import { ChartKind } from '@lightdash/common';
 import {
     ActionIcon,
     Group,
+    LoadingOverlay,
     ScrollArea,
     Stack,
     Title,
     Tooltip,
 } from '@mantine/core';
-import { IconLayoutSidebarLeftCollapse } from '@tabler/icons-react';
+import { IconLayoutSidebarLeftCollapse, IconReload } from '@tabler/icons-react';
 import { type Dispatch, type FC, type SetStateAction } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
 import { VisualizationConfigPanel } from '../../../components/DataViz/VisualizationConfigPanel';
+import { useRefreshTables } from '../hooks/useTables';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
 import { setSelectedChartType, SidebarTabs } from '../store/sqlRunnerSlice';
 import { TablesPanel } from './TablesPanel';
@@ -22,17 +24,41 @@ type Props = {
 export const Sidebar: FC<Props> = ({ setSidebarOpen }) => {
     const dispatch = useAppDispatch();
 
+    const projectUuid = useAppSelector((state) => state.sqlRunner.projectUuid);
+
+    const {
+        mutate: updateTables,
+        // TODO: unify this loading with the one in TablesPanel, handle error
+        isLoading,
+    } = useRefreshTables({ projectUuid });
+
     const { selectedChartType, activeSidebarTab, sqlColumns } = useAppSelector(
         (state) => state.sqlRunner,
     );
     return (
         <Stack spacing="xs" sx={{ flex: 1, overflow: 'hidden' }}>
             <Group position="apart">
-                <Title order={5} fz="sm" c="gray.6">
-                    {activeSidebarTab === SidebarTabs.TABLES
-                        ? 'TABLES'
-                        : 'VISUALIZATION'}
-                </Title>
+                <Group noWrap spacing="xs">
+                    <Title order={5} fz="sm" c="gray.6">
+                        {activeSidebarTab === SidebarTabs.TABLES
+                            ? 'TABLES'
+                            : 'VISUALIZATION'}
+                    </Title>
+                    {activeSidebarTab === SidebarTabs.TABLES && (
+                        <Tooltip
+                            variant="xs"
+                            label="Refresh tables"
+                            position="right"
+                        >
+                            <ActionIcon
+                                size="xs"
+                                onClick={() => updateTables()}
+                            >
+                                <MantineIcon icon={IconReload}></MantineIcon>
+                            </ActionIcon>
+                        </Tooltip>
+                    )}
+                </Group>
                 <Tooltip variant="xs" label="Close sidebar" position="left">
                     <ActionIcon size="xs">
                         <MantineIcon
@@ -49,6 +75,7 @@ export const Sidebar: FC<Props> = ({ setSidebarOpen }) => {
                 }
                 sx={{ flex: 1, overflow: 'hidden' }}
             >
+                <LoadingOverlay visible={isLoading} />
                 <TablesPanel />
             </Stack>