+++ packages/frontend/src/features/sqlRunner/store/selectors.ts
@@ -72,14 +72,14 @@ export const selectCurrentCartesianChartState = createSelector(
     },
 );
 
-const getXLayoutOptions = createSelector(
+const getIndexLayoutOptions = createSelector(
     [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.options?.xLayoutOptions,
+    (chartConfig) => chartConfig?.options?.indexLayoutOptions,
 );
 
-const getYLayoutOptions = createSelector(
+const getValuesLayoutOptions = createSelector(
     [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.options?.yLayoutOptions,
+    (chartConfig) => chartConfig?.options?.valuesLayoutOptions,
 );
 
 const getXAxisField = createSelector(
@@ -97,16 +97,16 @@ const getGroupByField = createSelector(
     (chartConfig) => chartConfig?.config?.fieldConfig?.groupBy?.[0],
 );
 
-const getGroupByLayoutOptions = createSelector(
+const getPivotLayoutOptions = createSelector(
     [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.options?.groupByOptions,
+    (chartConfig) => chartConfig?.options?.pivotLayoutOptions,
 );
 
 export const cartesianChartSelectors = {
-    getXLayoutOptions,
-    getYLayoutOptions,
+    getIndexLayoutOptions,
+    getValuesLayoutOptions,
     getXAxisField,
     getYAxisFields,
     getGroupByField,
-    getGroupByLayoutOptions,
+    getPivotLayoutOptions,
 };