+++ packages/warehouses/src/warehouseClients/PostgresWarehouseClient.ts
@@ -365,11 +365,13 @@ export class PostgresClient<
     async getAllTables() {
         const databaseName = this.config.database;
         const whereSql = databaseName ? `AND table_catalog = $1` : '';
+        const filterSystemTables = `AND table_schema NOT IN ('information_schema', 'pg_catalog')`;
         const query = `
             SELECT table_catalog, table_schema, table_name
             FROM information_schema.tables
             WHERE table_type = 'BASE TABLE'
                 ${whereSql}
+                ${filterSystemTables}
             ORDER BY 1, 2, 3
         `;
         const { rows } = await this.runQuery(